requirements.txt
ollama
openai
python-docx
ocr-tamil
pdf2image
indic-nlp-library
torch
groq